+++ modules/bookmark.py
@@ -1,0 +1,23 @@
+class BookmarkAction(QAction):
+    def __init__(self, title, parent):
+        super().__init__(title, parent)
+        self.setContextMenuPolicy(Qt.ActionsContextMenu)
+        self.customContextMenuRequested.connect(self.showContextMenu)
+        self.url = ""
+
+    def showContextMenu(self, point):
+        contextMenu = QMenu(self.parent())
+        deleteAction = QAction("削除", self)
+        deleteAction.triggered.connect(self.deleteBookmark)
+        contextMenu.addAction(deleteAction)
+        contextMenu.exec_(self.mapToGlobal(point))
+
+    def deleteBookmark(self):
+        tree = ET.parse('shortcuts.xml')
+        root = tree.getroot()
+        for shortcut in root.findall('shortcut'):
+            if shortcut.find('url').text == self.url:
+                root.remove(shortcut)
+                tree.write('shortcuts.xml')
+                break
+        self.parent().removeAction(self)
